[cli] Add configuration system for REPL to set all items per line (#208)
This PR adds a REPL configuration system so users can change calculator
display/computation options via `:...` meta-commands or inline
`expr:...` one-off overrides.

**Changes:**
- Introduce `Settings`, `parse_settings`, and `split_inline_settings`
for REPL configuration parsing.
- Integrate settings/meta-commands + inline overrides into the REPL
loop; add extensive unit tests for the settings parser.
- Expand docstrings across CLI calculator modules and update
planning/docs; adjust CLI delimiter flag (removes `-D`).

+++ tests/decimal128/test_decimal128_from_int.mojo
@@ -60,7 +60,7 @@ def test_from_int() raises:
     # from_int with scale: use a=integer, b=scale
     var scale_cases = load_test_cases(toml, "from_int_with_scale_tests")
     for tc in scale_cases:
-        var result = Dec128.from_int(Int(tc.a), Int(tc.b))
+        var result = Dec128.from_int(Int(tc.a), UInt32(Int(tc.b)))
         try:
             testing.assert_equal(
                 lhs=String(result),
@@ -218,7 +218,7 @@ def test_from_int_with_scale_advanced() raises:
     testing.assert_equal(r5.scale(), 25)
 
     # Max scale
-    var r6 = Dec128.from_int(1, Decimal128.MAX_SCALE)
+    var r6 = Dec128.from_int(1, UInt32(Decimal128.MAX_SCALE))
     testing.assert_equal(r6.scale(), Decimal128.MAX_SCALE)
 
     # Arithmetic with scaled value: 1.0 / 0.03
